refactor(config-ui): don't load custom hook when table is loading
