+++ config-ui/src/components/table/components/index.ts
@@ -18,3 +18,4 @@
 
 export * from './loading';
 export * from './no-data';
+export * from './content';
