+++ config-ui/src/components/table/table.tsx
@@ -17,40 +17,20 @@
  */
 
 import React from 'react';
-import { Checkbox, Radio } from '@blueprintjs/core';
 
-import { TextTooltip } from '@/components';
+import type { TableContentProps } from './components';
+import { TableLoading, TableNoData, TableContent } from './components';
 
-import { ColumnType } from './types';
-import { TableLoading, TableNoData } from './components';
-import { useRowSelection, UseRowSelectionProps } from './hooks';
-import * as S from './styled';
-
-interface Props<T> extends UseRowSelectionProps<T> {
+interface Props<T> extends TableContentProps<T> {
   loading?: boolean;
-  columns: ColumnType<T>;
-  dataSource: T[];
   noData?: {
     text?: React.ReactNode;
     btnText?: string;
     onCreate?: () => void;
   };
-  noShadow?: boolean;
 }
 
-export const Table = <T extends Record<string, any>>({
-  loading,
-  columns,
-  dataSource,
-  noData = {},
-  rowSelection,
-  noShadow = false,
-}: Props<T>) => {
-  const { canSelection, selectionType, getCheckedAll, onCheckedAll, getChecked, onChecked } = useRowSelection<T>({
-    dataSource,
-    rowSelection,
-  });
-
+export const Table = <T extends Record<string, any>>({ loading, dataSource, noData = {}, ...props }: Props<T>) => {
   if (loading) {
     return <TableLoading />;
   }
@@ -59,53 +39,5 @@ export const Table = <T extends Record<string, any>>({
     return <TableNoData {...noData} />;
   }
 
-  return (
-    <S.Table
-      style={{
-        boxShadow: noShadow ? 'none' : '0px 2.4px 4.8px -0.8px rgba(0, 0, 0, 0.1), 0px 1.6px 8px rgba(0, 0, 0, 0.07)',
-      }}
-    >
-      <S.THeader>
-        <S.TR>
-          {canSelection && (
-            <S.TH style={{ width: 40, textAlign: 'center' }}>
-              {selectionType === 'checkbox' && <Checkbox checked={getCheckedAll()} onChange={() => onCheckedAll()} />}
-            </S.TH>
-          )}
-          {columns.map(({ key, width, align = 'left', title }) => (
-            <S.TH key={key} style={{ width, textAlign: align }}>
-              {title}
-            </S.TH>
-          ))}
-        </S.TR>
-      </S.THeader>
-      <S.TBody>
-        {dataSource.map((data, i) => (
-          <S.TR key={i}>
-            {canSelection && (
-              <S.TD style={{ width: 40, textAlign: 'center' }}>
-                {selectionType === 'checkbox' && (
-                  <Checkbox checked={getChecked(data)} onChange={() => onChecked(data)} />
-                )}
-                {selectionType === 'radio' && <Radio checked={getChecked(data)} onChange={() => onChecked(data)} />}
-              </S.TD>
-            )}
-            {columns.map(({ key, width, align = 'left', ellipsis, dataIndex, render }) => {
-              const value = Array.isArray(dataIndex)
-                ? dataIndex.reduce((acc, cur) => {
-                    acc[cur] = data[cur];
-                    return acc;
-                  }, {} as any)
-                : data[dataIndex];
-              return (
-                <S.TD key={key} style={{ width, textAlign: align }}>
-                  {render ? render(value, data) : ellipsis ? <TextTooltip content={value}>{value}</TextTooltip> : value}
-                </S.TD>
-              );
-            })}
-          </S.TR>
-        ))}
-      </S.TBody>
-    </S.Table>
-  );
+  return <TableContent dataSource={dataSource} {...props} />;
 };
